% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basis_label}
\alias{basis_label}
\title{Return the display label for a basis.
Used by plot and summary methods.}
\usage{
basis_label(basis_name)
}
\description{
Return the display label for a basis.
Used by plot and summary methods.
}
\keyword{internal}
