% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{compute_basis_matrix}
\alias{compute_basis_matrix}
\title{Compute the density or cumulative basis matrix.
Replaces \code{basis_mpl()}.}
\usage{
compute_basis_matrix(x, knots, basis_name, order, which = 1)
}
\description{
Compute the density or cumulative basis matrix.
Replaces \code{basis_mpl()}.
}
\keyword{internal}
