% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{compute_knots}
\alias{compute_knots}
\title{Compute the knot sequence for a given basis.
Replaces \code{knots_mpl()}.}
\usage{
compute_knots(control, events)
}
\description{
Compute the knot sequence for a given basis.
Replaces \code{knots_mpl()}.
}
\keyword{internal}
