% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{compute_penalty}
\alias{compute_penalty}
\title{Compute the roughness penalty matrix.
Replaces \code{penalty_mpl()}.}
\usage{
compute_penalty(control, knots)
}
\description{
Compute the roughness penalty matrix.
Replaces \code{penalty_mpl()}.
}
\keyword{internal}
