% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{compute_penalty_order}
\alias{compute_penalty_order}
\title{Compute the penalty order for a basis.
Replaces \code{penalty.order_mpl()}.}
\usage{
compute_penalty_order(basis_name, p, order)
}
\description{
Compute the penalty order for a basis.
Replaces \code{penalty.order_mpl()}.
}
\keyword{internal}
